+++ src/utils/index.ts
@@ -1,4 +1,8 @@
+export * from './AudioPlayer';
+export { default as AudioPlayer } from './AudioPlayer';
+export * from './base64';
 export * from './color';
-export * from './math';
-export * from './string';
 export * from './crypto';
+export * from './math';
+export * from './stream';
+export * from './string';